+++ CHANGELOG.md
@@ -39,6 +39,7 @@
 - Sale Order : Hide 'Order being edited' flag when complete manually edited sale order.
 - FixedAsset : Generation / Prorata. Change computation process for Linear method when isProrataTemporis=true
 - Stock Location Line: Future quantity is now computed from a query.
+- Stock Location Line: Requested reserved quantity is now computed from a query.
 - ACCOUNT CONFIG : Remove fields allowNonExportedMoveEditing and supplierAccount.Reorganize form view.
 - PROJECT : Add mass update for "imputable" field